fix: use uint64 in SslFactory call in sslCmdRepo

+++ src/infra/sslCmdRepo.go
@@ -52,7 +52,7 @@ func (repo SslCmdRepo) Add(addSsl dto.AddSsl) error {
 	}
 
 	newSsl, err := sslQueryRepo.SslFactory(
-		int(sslId.Get()),
+		uint64(sslId.Get()),
 		addSsl.Hostname.String(),
 		addSsl.Key.String(),
 		addSsl.Certificate.String(),
